Cash and Atm implementations were divided
