+++ src/main/java/com/solvd/atm/MainClass.java
@@ -3,8 +3,10 @@ package com.solvd.atm;
 import com.solvd.atm.domain.*;
 import com.solvd.atm.service.AtmService;
 import com.solvd.atm.service.BankService;
+import com.solvd.atm.service.CashService;
 import com.solvd.atm.service.impl.AtmServiceImpl;
 import com.solvd.atm.service.impl.BankServiceImpl;
+import com.solvd.atm.service.impl.CashServiceImpl;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -161,8 +163,9 @@ public class MainClass {
 //        CardService cardService = new CardServiceImpl();
 //        System.out.println(cardService.read(3L));
 
-        AtmService atmService = new AtmServiceImpl();
-        System.out.println(atmService.getMap());
+//        AtmService atmService = new AtmServiceImpl();
+        CashService cs = new CashServiceImpl();
+        System.out.println(cs.getMap());
 
 //        ClientService clientService = new ClientServiceImpl();
 //        clientService.update(client1);
