+++ src/main/java/com/solvd/atm/persistence/AtmRepository.java
@@ -14,8 +14,6 @@ public interface AtmRepository {
 
     Optional<Atm> read(Long id);
 
-    Map<CurrencyType, Map<BigDecimal, BigDecimal>> getMap();
-
     void update(Atm atm);
 
     void deleteById(Long deleteId);
