+++ src/main/java/com/solvd/atm/persistence/CashRepository.java
@@ -1,0 +1,12 @@
+package com.solvd.atm.persistence;
+
+import com.solvd.atm.domain.*;
+
+import java.math.BigDecimal;
+import java.util.Map;
+
+public interface CashRepository {
+
+    Map<CurrencyType, Map<BigDecimal, BigDecimal>> getMap();
+
+}
