+++ src/main/java/com/solvd/atm/persistence/impl/AtmMapperImpl.java
@@ -1,17 +1,10 @@
 package com.solvd.atm.persistence.impl;
 
 import com.solvd.atm.domain.Atm;
-import com.solvd.atm.domain.Cash;
-import com.solvd.atm.domain.CurrencyType;
 import com.solvd.atm.persistence.AtmRepository;
 import com.solvd.atm.persistence.MyBatisConfig;
-import org.apache.ibatis.session.ResultContext;
-import org.apache.ibatis.session.ResultHandler;
 import org.apache.ibatis.session.SqlSession;
 
-import java.math.BigDecimal;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Optional;
 
 public class AtmMapperImpl implements AtmRepository {
@@ -29,27 +22,6 @@ public class AtmMapperImpl implements AtmRepository {
         try (SqlSession sqlSession = MyBatisConfig.getSqlSessionFactory().openSession(true)) {
             AtmRepository atmRepository = sqlSession.getMapper(AtmRepository.class);
             return atmRepository.read(id);
-        }
-    }
-
-    public Map<CurrencyType, Map<BigDecimal, BigDecimal>> getMap() {
-        try (SqlSession sqlSession = MyBatisConfig.getSqlSessionFactory().openSession(true)) {
-            class AtmHandler implements ResultHandler {
-                final Map<CurrencyType, Map<BigDecimal, BigDecimal>> atmCash = new HashMap<CurrencyType, Map<BigDecimal, BigDecimal>>();
-
-                @Override
-                public void handleResult(ResultContext context) {
-                    final Cash complex = (Cash)context.getResultObject();
-                    if (!atmCash.containsKey(complex.getCurrencyType())) {
-                        atmCash.put(complex.getCurrencyType(), new HashMap<BigDecimal, BigDecimal>());
-                    }
-                    atmCash.get(complex.getCurrencyType()).put(complex.getDenomination(), complex.getQuantity());
-                }
-            }
-            AtmHandler handler = new AtmHandler();
-            sqlSession.select("com.solvd.atm.persistence.AtmRepository.getMap", null, handler);
-
-            return handler.atmCash;
         }
     }
