+++ src/main/java/com/solvd/atm/service/AtmService.java
@@ -13,8 +13,6 @@ public interface AtmService {
 
     Optional<Atm> read(Long id);
 
-    Map<CurrencyType, Map<BigDecimal, BigDecimal>> getMap();
-
     void update(Atm atm);
 
     void deleteById(Long deleteId);
