+++ src/main/java/com/solvd/atm/service/impl/AtmServiceImpl.java
@@ -1,13 +1,10 @@
 package com.solvd.atm.service.impl;
 
 import com.solvd.atm.domain.Atm;
-import com.solvd.atm.domain.CurrencyType;
 import com.solvd.atm.persistence.AtmRepository;
 import com.solvd.atm.persistence.impl.AtmMapperImpl;
 import com.solvd.atm.service.AtmService;
 
-import java.math.BigDecimal;
-import java.util.Map;
 import java.util.Optional;
 
 public class AtmServiceImpl implements AtmService {
@@ -28,11 +25,6 @@ public class AtmServiceImpl implements AtmService {
     @Override
     public Optional<Atm> read(Long id) {
         return atmRepository.read(id);
-    }
-
-    @Override
-    public Map<CurrencyType, Map<BigDecimal, BigDecimal>> getMap() {
-        return atmRepository.getMap();
     }
 
     @Override
