+++ src/main/resources/config.properties
@@ -1,5 +1,5 @@
 driver=com.mysql.cj.jdbc.Driver
 url=jdbc:mysql://localhost:3306/laba_atm
 userName=root
-password=Vfhbz011288
+password=Ant1d%te
 poolSize=5
