+++ src/main/resources/mapper/AtmMapper.xml
@@ -2,30 +2,19 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace ="com.solvd.atm.persistence.AtmRepository">
     <insert id="create" keyColumn="id" keyProperty="atm.id" useGeneratedKeys="true">
-        insert into atms (bank_id, address_id, cash_id)
-        values #{bankId}, #{addressId}, #{cashId}}
+        insert into atms (bank_id, address_id)
+        values #{bankId}, #{addressId}
     </insert>
 
     <select id="read" resultMap="AtmResultMap">
-        select id, bank_id, address_id, cash_id from atms where id = #{id}
-    </select>
-
-    <select id="getMap" resultMap="CashResultMap">
-        select currency_type, denomination, quantity from cash
+        select id, bank_id, address_id from atms where id = #{id}
     </select>
 
     <resultMap id="AtmResultMap" type="com.solvd.atm.domain.Atm" autoMapping="false">
         <id property="id" column="id"/>
-        <association property= "balance" resultMap="com.solvd.atm.persistence.AtmRepository.CashResultMap"/>
+        <association property= "balance" resultMap="com.solvd.atm.persistence.CashRepository.CashResultMap"/>
         <association property="bank" resultMap="com.solvd.atm.persistence.BankRepository.BankResultMap"/>
         <association property="address" resultMap="com.solvd.atm.persistence.AddressRepository.AddressResultMap"/>
-    </resultMap>
-
-    <resultMap id="CashResultMap" type="com.solvd.atm.domain.Cash" autoMapping="false">
-        <id property="id" column="id"/>
-        <result property= "currencyType" column="currency_type"/>
-        <result property= "denomination" column="denomination"/>
-        <result property= "quantity" column="quantity"/>
     </resultMap>
 
     <update id="update" useGeneratedKeys="false">
