+++ src/main/resources/mapper/CashMapper.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace ="com.solvd.atm.persistence.CashRepository">
+    <select id="getMap" resultMap="CashResultMap">
+        select currency_type, denomination, quantity from cash
+    </select>
+
+    <resultMap id="CashResultMap" type="com.solvd.atm.domain.Cash" autoMapping="false">
+        <id property="id" column="id"/>
+        <result property= "currencyType" column="currency_type"/>
+        <result property= "denomination" column="denomination"/>
+        <result property= "quantity" column="quantity"/>
+    </resultMap>
+</mapper>
